auth

+++ dia4/MERN_ECOMMERCE/backend/src/routes/user.route.js
@@ -1,7 +1,7 @@
 const {Router} = require('express')
 const router = Router()
 
-const {create,getAll,getOne,updateOne,deleteOne} = require('../controllers/user.controller')
+const {create,getAll,getOne,updateOne,deleteOne, auth} = require('../controllers/user.controller')
 
 router.route('/')
     .post(create)
@@ -11,5 +11,8 @@ router.route('/:id')
     .get(getOne)
     .put(updateOne)
     .delete(deleteOne)
+
+router.route('/auth')
+    .post(auth)
 
 module.exports = router
